fix config

+++ p2/main.cpp
@@ -20,7 +20,7 @@ public:
 
         int ans = 0;
 
-        for (int i = 0; i < s.length()::int ; i++)
+        for (int i = 0; i < s.length(); i++)
         {
             if (m[s[i]] < m[s[i + 1]])
             {
